% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accel}
\alias{accel}
\title{Acceleration}
\format{
A \code{"data.frame"} structure with 1732 observations, 3 nominal and 11 numerical predictor variables
}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/bike+sharing+dataset}{UCI Archive}
}
\usage{
data(accel)
}
\description{
Dataset with acceleration target value w.r.t. 14 nominal and numerical variables
}
\examples{
data(accel)
head(accel)
}
\references{
Hadi Fanaee-T. and João Gama. Event labeling combining ensemble detectors and background knowledge. Prog. in Art. Int., pages 1-15, 2013. ISSN 2192-6352. (\href{https://link.springer.com/article/10.1007/s13748-013-0040-3}{Springer})
}
\keyword{datasets}
