% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdMetrics.R
\name{corr}
\alias{corr}
\title{Pearson's Correlation}
\usage{
corr(trues, preds)
}
\arguments{
\item{trues}{Target values from a test set of a given data set. Should be a vector and have the same size as the variable preds}

\item{preds}{Predicted values given a certain test set of a given data set. Should be a vector and have the same size as the variable trues}
}
\value{
Value for the Pearson's correlation
}
\description{
Pearson's Correlation
}
\examples{
library(IRon)
library(earth)

data(accel)

form <- acceleration ~ .

ind <- sample(1:nrow(accel),0.75*nrow(accel))

train <- accel[ind,]
test <- accel[-ind,]

m <- earth::earth(form, train)
preds <- as.vector(predict(m,test))

trues <- test$acceleration

corr(trues,preds)
}
