% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_plot.R
\name{errorPlot}
\alias{errorPlot}
\title{Visualization of comparison between prediction error and item's relevance}
\usage{
errorPlot(trues, preds, phi.parms, e = 1, thr = 0)
}
\arguments{
\item{trues}{True target values of a given test set}

\item{preds}{Predicted values for the test set}

\item{phi.parms}{The relevance function providing the data points where the pairs of values-relevance are known (use ?phi.control() for more information)}

\item{e}{Exponent for distance - 1 is linear (mae) and 2 is quadratic (mse)}

\item{thr}{Relevance threshold (default 0)}
}
\value{
Produces a plot with a visualization comparing prediction error and item's relevance
}
\description{
The errorPlot function takes a set of target and predicted values to calculate the deviation as a function of phi and creates a plot with points and an approximate function err(phi).
}
\examples{
\dontrun{
library(IRon)
require(earth)
require(mgcv)

data(accel)

ind <- sample(1:nrow(accel),0.75*nrow(accel))

train <- accel[ind,]
test <- accel[-ind,]

ph <- phi.control(accel$acceleration)

m <- earth::earth(acceleration ~ ., train)
preds <- as.vector(predict(m,test))

trues <- accel$acceleration

errorPlot(trues,preds,phi.parms = ph)
}


}
