% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{fit}
\alias{fit}
\alias{quant_reg}
\alias{res_quant}
\alias{cooks_dist}
\alias{envelope}
\title{xx}
\usage{
quant_reg(
  formula,
  link = "identity",
  family = "normal",
  tau = 0.5,
  data = NULL,
  out.list = FALSE,
  estimate.df = FALSE,
  df = 4,
  ...
)

res_quant(model)

cooks_dist(model, npoints = 0, plot = FALSE, data, cex = 1, ...)

envelope(
  model,
  k = 100,
  color = "grey50",
  xlabel = "Theorical Quantile",
  ylabel = "Empirical Quantile",
  font = "serif",
  cex.axis = 1,
  cex.lab = 1
)
}
\arguments{
\item{formula}{x}

\item{link}{x}

\item{family}{x}

\item{tau}{x}

\item{data}{x}

\item{out.list}{x}

\item{estimate.df}{x}

\item{df}{x}

\item{...}{x}

\item{model}{x}

\item{npoints}{x}

\item{plot}{x}

\item{cex}{x}

\item{k}{x}

\item{color}{x}

\item{xlabel}{x}

\item{ylabel}{x}

\item{font}{x}

\item{cex.axis}{x}

\item{cex.lab}{x}
}
\description{
XX
}
\examples{

set.seed(2)
y <- VGAM::rbisa(100, shape = 4, scale = 1)
quant_reg(y ~ 1, link = 'log', family = 'normal', tau = 0.9)
quant_reg(y ~ 1, link = 'log', family = 'pe', tau = 0.5)
quant_reg(y ~ 1, link = 'log', family = 't', tau = 0.5)
quant_reg(y ~ 1, link = 'log', family = 'logis', tau = 0.5)
quant_reg(y ~ 1, link = 'log', family = 'pe', tau = 0.2)
quant_reg(y ~ 1, link = 'log', family = 't', tau = 0.8)

}
