% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdMetrics.R
\name{mae}
\alias{mae}
\title{Standard Evaluation Metrics}
\usage{
mae(trues, preds)
}
\arguments{
\item{trues}{Target values from a test set of a given data set. Should be a vector and have the same size as the variable preds}

\item{preds}{Predicted values given a certain test set of a given data set. Should be a vector and have the same size as the variable trues}
}
\value{
Value for the mean average error
}
\description{
Mean Average Error
}
\keyword{internal}
