% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi.range}
\alias{phi.range}
\title{Custom Relevance Function}
\usage{
phi.range(y, control.pts, ...)
}
\arguments{
\item{y}{The target variable of a given data set}

\item{control.pts}{Parameter representing a 3-column matrix of y-value, corresponding relevance value (between 0 and 1), and the derivative of such relevance value, allowing users to specify the known relevance at given target values}

\item{...}{Additional parameters}
}
\value{
A list with three slots with information concerning the relevance function
\item{method}{The method used to generate the relevance function (extremes or range)}
\item{npts}{?}
\item{control.pts}{Three sets of values identifying the target value-relevance-derivate for the first low extreme value, the median, and first high extreme value}
}
\description{
User-guided approach to obtain a relevance function for certain intervals of the target variable when the option of range is chosen in function phi.control, i.e. users define the relevance of values for which it is known
}
\keyword{internal}
