% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdMetrics.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(trues, preds)
}
\arguments{
\item{trues}{Target values from a test set of a given data set. Should be a vector and have the same size as the variable preds}

\item{preds}{Predicted values given a certain test set of a given data set. Should be a vector and have the same size as the variable trues}
}
\value{
Value for the relevance-weighted root mean squared error
}
\description{
Root Mean Squared Error
}
\keyword{internal}
