% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdMetrics.R
\name{variance}
\alias{variance}
\title{Model Variance}
\usage{
variance(preds)
}
\arguments{
\item{preds}{Predicted values given a certain test set of a given data set. Should be a vector and have the same size as the variable trues}
}
\value{
Value for model variance
}
\description{
Model Variance
}
\keyword{internal}
