% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{fit}
\alias{fit}
\alias{quant_reg}
\alias{res_quant}
\alias{cooks_dist}
\alias{envelope}
\title{Fitting Quantile Regression Models}
\usage{
quant_reg(
  formula,
  link = "identity",
  family = "normal",
  tau = 0.5,
  data = NULL,
  out.list = FALSE,
  estimate.df = FALSE,
  df = 4,
  ...
)

res_quant(model)

cooks_dist(model, npoints = 0, plot = FALSE, data, cex = 1, ...)

envelope(
  model,
  k = 100,
  color = "grey50",
  xlabel = "Theorical Quantile",
  ylabel = "Empirical Quantile",
  font = "serif",
  cex.axis = 1,
  cex.lab = 1
)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{link}{link function to be used in the model.}

\item{family}{a description of the kernel.}

\item{tau}{quantile.}

\item{data}{an optional data frame containing the variables of the model.}

\item{out.list}{logical.}

\item{estimate.df}{logical.}

\item{df}{degrees of freedom.}

\item{...}{additional arguments to be passed.}

\item{model}{object of class iron.}

\item{npoints}{number of points to be identified.}

\item{plot}{logical.}

\item{cex}{a numerical vector giving the amount by which plotting characters and symbols should be scaled relative to the default.}

\item{k}{number of replications for envelope construction.}

\item{color}{a specification for the default envelope color.}

\item{xlabel}{a label for the x axis.}

\item{ylabel}{a label for the y axis.}

\item{font}{the name of a font family for x and y axis.}

\item{cex.axis}{The magnification to be used for axis annotation relative to the current setting of cex.}

\item{cex.lab}{The magnification to be used for x and y labels relative to the current setting of cex.}
}
\description{
quant_reg is used to fit quantile regression models, specified by giving a symbolic description of the linear predictor and a description of the kernel.
}
\examples{

set.seed(2)
y <- VGAM::rbisa(100, shape = 4, scale = 1)
quant_reg(y ~ 1, link = 'log', family = 'normal', tau = 0.9)
quant_reg(y ~ 1, link = 'log', family = 'pe', tau = 0.5)
quant_reg(y ~ 1, link = 'log', family = 't', tau = 0.5)
quant_reg(y ~ 1, link = 'log', family = 'logis', tau = 0.5)
quant_reg(y ~ 1, link = 'log', family = 'pe', tau = 0.2)
quant_reg(y ~ 1, link = 'log', family = 't', tau = 0.8)

}
