% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdMetrics.R
\name{variance}
\alias{variance}
\title{Model Variance}
\usage{
variance(preds)
}
\arguments{
\item{preds}{Predicted values given a certain test set of a given data set. Should be a vector and have the same size as the variable trues}
}
\value{
Value for model variance
}
\description{
Model Variance
}
\examples{
library(IRon)
library(earth)

data(accel)

form <- acceleration ~ .

ind <- sample(1:nrow(accel),0.75*nrow(accel))

train <- accel[ind,]
test <- accel[-ind,]

m <- earth::earth(form, train)
preds <- as.vector(predict(m,test))

trues <- test$acceleration

variance(preds)
}
